Refactor FileField handling: use field storage/upload_to, avoid open file handles
- Extract file logic into files.py; use field.storage instead of default_storage
- Use field.generate_filename() so upload_to callables are respected
- __getitem__ returns raw storage path string instead of an opened File handle
- set_item only saves on actual UploadedFile, skips no-op on existing value
